refactor(163): reorganize reusable ui components by family
