+++ app/src/main/java/org/cescfe/numpairs/feature/game/ui/GameScreenSections.kt
@@ -32,10 +32,10 @@ import org.cescfe.numpairs.feature.game.presentation.StripItemUiState
 import org.cescfe.numpairs.feature.game.presentation.StripItemVisualStyle
 import org.cescfe.numpairs.feature.game.presentation.TileOperatorSelectionDialogUiState
 import org.cescfe.numpairs.feature.game.presentation.TileUiState
-import org.cescfe.numpairs.feature.game.ui.components.AvailableNumberChip
-import org.cescfe.numpairs.feature.game.ui.components.AvailableNumberChipStyle
-import org.cescfe.numpairs.feature.game.ui.components.AvailableNumberInputChip
-import org.cescfe.numpairs.feature.game.ui.components.PuzzleTile
+import org.cescfe.numpairs.feature.game.ui.components.strip.AvailableNumberChip
+import org.cescfe.numpairs.feature.game.ui.components.strip.AvailableNumberChipStyle
+import org.cescfe.numpairs.feature.game.ui.components.strip.AvailableNumberInputChip
+import org.cescfe.numpairs.feature.game.ui.components.tile.PuzzleTile
 import org.cescfe.numpairs.ui.theme.NumPairsComponents
 
 @Composable
